2024-28/Face Recognition based login system/pyproject.toml
[project]
name = "face-recognition-based-login-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "deepface>=0.0.95",
    "flask>=3.1.2",
    "flask-cors>=6.0.1",
    "flask-jwt-extended>=4.7.1",
    "opencv-python>=4.12.0.88",
    "python-dotenv>=1.1.1",
    "tf-keras>=2.20.1",
]
